requirements.txt
alembic==0.8.10
apache-airflow==1.9.0
appnope==0.1.0
argcomplete==1.8.2
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.5.3
bleach==2.1.2
brewer2mpl==1.4.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorlover==0.2.1
configparser==3.5.0
croniter==0.3.23
crypto==1.4.1
cryptography==2.2.2
cufflinks==0.12.1
cycler==0.10.0
decorator==4.2.1
dill==0.2.7.1
docutils==0.14
docx2txt==0.6
EbookLib==0.15
entrypoints==0.2.3
Flask==0.11.1
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.14
funcsigs==1.0.0
future==0.16.0
ggplot==0.11.5
gitdb2==2.0.3
GitPython==2.1.10
gunicorn==19.8.1
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.8.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
kiwisolver==1.0.1
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
Naked==0.1.31
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.2
ordereddict==1.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
patsy==0.5.0
pdfminer.six==20170720
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
plotly==2.5.1
pocketsphinx==0.1.3
prompt-toolkit==1.0.15
psutil==4.4.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycparser==2.18
pycryptodome==3.6.1
Pygments==2.2.0
pymongo==3.6.1
pyparsing==2.2.0
PyPDF2==1.26.0
python-daemon==2.1.2
python-dateutil==2.7.0
python-editor==1.0.3
python-nvd3==0.14.2
python-pptx==0.6.5
python-slugify==1.1.4
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scipy==1.0.1
Send2Trash==1.5.0
setproctitle==1.1.10
shellescape==3.4.1
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.3
SpeechRecognition==3.6.3
SQLAlchemy==1.2.5
statsmodels==0.8.0
tabulate==0.7.7
terminado==0.8.1
testpath==0.3.1
textract==1.6.1
thrift==0.11.0
tornado==5.0
traitlets==4.3.2
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
WTForms==2.1
xlrd==1.0.0
XlsxWriter==1.0.4
xmltodict==0.11.0
zope.deprecation==4.3.0
